Don't skip the preparation while starting a selection over an existing highlight.
This partially reverts 42909c058c9c3865f6254dfae03bdb2c883dc8e0.

Some explanation: since the selection activity is started, independently of the result if the (now disabled) check for an existing highlight (or other annotator-specific element),
it is not consitent for us to behave differently.

Here is an example. Create two annotations in a text, like this:

"whatever [long] text [here], and more."

If you start a selection inside the work "text", and move into the word "long", the viewer (with the annotation for "long") will not show up. That's intended.
However, if you start the same selection inside the "here" word, which already has an annotation, and then move into the "long" word, the viewer for long pops up, since isMouseDown was not set, because of this test.

I think this is not what we want, so I removed this test case.

For protection agains annotating ourselves, we have other measures in checkForEndSelection.