use old getSelectedRanges; add getTargetFromRange
With this change the old getSelectedRanges is identical to before the
fuzzy anchoring. In this way, a minimum of work is done when the
selection ends. A new getTargetFromRange method is added to the
annotator and used during setupAnnotation when the annotation has no
target already. This is where the alternative selectors are extracted.

This also fixes the problem I introduced where only the range selector
was getting stored.