Restore the top-level function wrapper in the concatenated build files.
Switched to slightly different toolchain for building the concatenated
JS files. The earlier version has causesd some problems, because they
lost the top-level function wrapper they used to have.

The challenge here was to have source maps that point to the original
coffee files, and a concatenated JS file which contains all the code,
with a single top-level function wrapper. And a preamble.

The solution was to introduce (a fork of) [coffeebar](https://npmjs.org/package/coffeebarx),
which is capable of concatenating multiple source files, adding a top-level
wrapper, and generating source maps the point back to the original files.

This, together with mapcat and uglifyjs can do what we want.