ck--/ann#67 fix the namespace problem
the version of annotator we use used the -b option (bare) option to
compile
coffeescript. This means:
  Compile the JavaScript without the top-level function safety wrapper
It essentially makes all variables global! Normally coffeescript has
a very stric module concept and this option puts all files into
the global namespace.

No idea why (our branch of) annotator does this, but this was a cause
of many errors.