Rip dom-text-mapper out of core.
Previously, dom-text-mapper was a required part of Annotator,
even when the TextAnchors plugin (which is actually using it)
was not loaded.

That was because it has also been providing the (fake) pagination
support, which is required for other anchor types, too.

Now I have ripped out the (fake stub) pagination support to a new
class, DummyDocumentAccess, and now that's the default
document access strategy shipped with Annotator.

There is a new plugin, called DomTextMapper, which actually registers
dom-text-mapper as a more sophisticated document access strategy,
which can provide all the text extraction and mapping facilities
required by the TextAnchors plugin. (So the TextAnchors plugin
depends on this plugin.)