Separate the TextHighlight plugin from TextAnchors plugin.
I have done this in preparation of creating a new, simplified version
of the TextAnchors plugin, which does not care about TextPosition
or TextQuote selectors; so, in short, provides only the functionality
that upstream Annotator provides currently - but in the new framework.

The benefit is that this new upcoming simplified text anchors plugin
does not need to depend on the dom-text-mapper plugin, which is required
for the advenced functionality.