Re-unite the TextAnchors and the OldTextAnchors plugin.
OldTextAnchor was born out of the desire to provide text anchoring
functions without creating a dependency the DomTextMapper plugin.

Now it's functions have been integrated back to the TextAnchors plugin,
and the dependency on DomTextMapper has been softened into a suggestion.

It goes like this:
 * When the TextAnchors plugin is initialized, it checks the presence of
   the DomTextMapper plugin, and sets a flag accordingly.
 * When the functions of the TextAnchors plugin are accessed, they
   observe this @useDTM flag, and act accordingly.

The differences are as follows:

When using the DomTextMapper plugin,
 * We can save more selectors.
   (TextPositionSelector, and a richer TextQuoteSelector)
 * We can use more anchoring strategies
   (The one based the TextPoisionSelector, and the two fuzzy strategies,
   from the FuzzyTextAnchors plugin)
 * We determine the quote of the annotation using DomTextMapper,
   which provides better results in various circumstances.
   (Whitespacing is closer to ideal, invisible content is omitted, etc.)

When not using the DomTextMapper plugin, text anchoring will still work,
but with less features.