dev.html: Change config to allow using the DTM doc access policy
Because we were already loading the DomTextMapper plugin, the only think
necessary for this was to pass in the "noScan" parameter, which delays
the scanning of the document (and with it, the initialization of the
document access policy) from annotator constructor, to the time when we
actually whant to anchor some annotations.

This way, the plugins are initialized first, and they can register the
more sophisticated document access strategy.