Removed the `noDocAccess` option
This was a really weird thing to add in the first place.
Who, in his right mind, would want to create an Annotator instance
that does not access the document? That is the whole point!

(Of course I know the answer: we, in the H project. But those specialized
use cases should be implemented by using overrides, and not by pushing
these options into upstream.)