Introducing _clientdeleteonly property for Store plugin
Sometimes we need to delete annotations just from the client side and we do not want our system to delete the
annotation from the backend. However the store plugin listened to the annotationDeleted event and automatically
deleted the annotation from the backend. Now, annotations can be tagged with the _clientdeleteonly property in
which case the store plugin only unregisters the annotation and does not trigger a delete to the backend.